sync_db_lambda/requirements.txt
instructure-dap-client[postgresql]==1.1.0
aws-lambda-powertools==2.34.2
boto3